requirements.txt
numpy~=1.22.3
matplotlib~=3.4.3
keras~=2.8.0
Flask~=2.1.1
Werkzeug~=2.1.1
gunicorn~=20.1.0

opencv-python~=4.5.5.64
mtcnn~=0.1.1
Pillow~=9.1.0
